requirements.txt
aiosqlite==0.22.1
httpx==0.28.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
python-dotenv>=1.0.0
